+++ modules/flopy/modflow/mfpar.py
@@ -221,13 +221,13 @@ class ModflowPar(object):
                 partyp = t[1].lower()
                 if partyp not in par_types:
                     par_types.append(partyp)
-                parval = np.float(t[2])
-                nclu = np.int(t[3])
+                parval = np.float64(t[2])
+                nclu = np.int_(t[3])
                 clusters = []
                 for nc in range(nclu):
                     line = f.readline()
                     t = line.strip().split()
-                    lay = np.int(t[0])
+                    lay = np.int_(t[0])
                     s = t[1]
                     if len(s) > 10:
                         s = s[0:10]
@@ -239,7 +239,7 @@ class ModflowPar(object):
                     iarr = []
                     for iv in t[3:]:
                         try:
-                            iz = int(np.int(iv))
+                            iz = int(np.int_(iv))
                             if iz != 0:
                                 iarr.append(iz)
                         except:
@@ -291,18 +291,18 @@ class ModflowPar(object):
 
 
         """
-        dtype = np.float32
+        dtype = np.float64
         data = np.zeros(shape, dtype=dtype)
         for key, tdict in parm_dict.items():
             partyp, parval = tdict['partyp'], tdict['parval']
             nclu, clusters = tdict['nclu'], tdict['clusters']
             if model.mfpar.pval is None:
-                pv = np.float(parval)
+                pv = np.float64(parval)
             else:
                 try:
-                    pv = np.float(model.mfpar.pval.pval_dict[key.lower()])
+                    pv = np.float64(model.mfpar.pval.pval_dict[key.lower()])
                 except:
-                    pv = np.float(parval)
+                    pv = np.float64(parval)
             # print partyp, parval, nclu, clusters
             if partyp == findkey:
                 for [layer, mltarr, zonarr, izones] in clusters:
