+++ modules/flopy/modflow/mfriv.py
@@ -238,18 +238,18 @@ class ModflowRiv(Package):
         # get an empty recarray that corresponds to dtype
         dtype = ModflowRiv.get_default_dtype(structured=structured)
         if aux_names is not None:
-            dtype = Package.add_to_dtype(dtype, aux_names, np.float32)
+            dtype = Package.add_to_dtype(dtype, aux_names, np.float64)
         return create_empty_recarray(ncells, dtype, default_value=-1.0E+10)
 
     @staticmethod
     def get_default_dtype(structured=True):
         if structured:
-            dtype = np.dtype([("k", np.int), ("i", np.int),
-                              ("j", np.int), ("stage", np.float32),
-                              ("cond", np.float32), ("rbot", np.float32)])
+            dtype = np.dtype([("k", np.int_), ("i", np.int_),
+                              ("j", np.int_), ("stage", np.float64),
+                              ("cond", np.float64), ("rbot", np.float64)])
         else:
-            dtype = np.dtype([("node", np.int), ("stage", np.float32),
-                              ("cond", np.float32), ("rbot", np.float32)])
+            dtype = np.dtype([("node", np.int_), ("stage", np.float64),
+                              ("cond", np.float64), ("rbot", np.float64)])
 
         return dtype
