+++ modules/flopy/pakbase.py
@@ -119,7 +119,7 @@ class Package(PackageInterface):
         self.url = 'index.html'
         self.allowDuplicates = allowDuplicates
 
-        self.acceptable_dtypes = [int, np.float32, str]
+        self.acceptable_dtypes = [int, np.float64, str]
 
         return
 
@@ -703,10 +703,10 @@ class Package(PackageInterface):
         nppak = 0
         if "parameter" in line.lower():
             t = line.strip().split()
-            nppak = np.int(t[1])
+            nppak = np.int_(t[1])
             mxl = 0
             if nppak > 0:
-                mxl = np.int(t[2])
+                mxl = np.int_(t[2])
                 if model.verbose:
                     msg = 3 * ' ' + 'Parameters detected. Number of ' + \
                           'parameters = {}'.format(nppak)
@@ -732,7 +732,7 @@ class Package(PackageInterface):
                     msg = 3 * ' ' + 'implicit nppak in {}'.format(filename)
                     print(msg)
             if nppak > 0:
-                mxl = np.int(t[3])
+                mxl = np.int_(t[3])
                 imax += 1
                 if model.verbose:
                     msg = 3 * ' ' + 'Parameters detected. Number of ' + \
@@ -871,12 +871,12 @@ class Package(PackageInterface):
 
                 #  get appropriate parval
                 if model.mfpar.pval is None:
-                    parval = np.float(par_dict['parval'])
+                    parval = np.float64(par_dict['parval'])
                 else:
                     try:
-                        parval = np.float(model.mfpar.pval.pval_dict[pname])
+                        parval = np.float64(model.mfpar.pval.pval_dict[pname])
                     except:
-                        parval = np.float(par_dict['parval'])
+                        parval = np.float64(par_dict['parval'])
 
                 # fill current parameter data (par_current)
                 for ibnd, t in enumerate(data_dict):
